+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.23.0
  - fix: `NesDialog` would rely on the "root" theme, which would break if it was not a nes ui one.
+ - feat: Dialogs now can be "confirmed" with the enter key, and the cancel (esc key) can be intercepted by the developer now
  - feat: adding `NesProgressBar` and `NesProgressBarTheme`.
 
 # 0.22.1